requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
flask-migrate